refactor: move importVal command to commands
